// src/app/models/rworkflow-model.ts
export class RWorkFlows {
  Id: number;
  Name: string;
  UILabel: string;
  CRAllowed: boolean;
  CRWFName: string;
  EntityType: string;
  EntityId: number;
  DatabaseId: number;
  CreatedById: number;
  UpdatedById: number;
  ErrorMessage: string;
  Description: number;
  ParameterCarrier: number;
  DiagramJson: string;
  Status: string;
  HelpFilePath: string;
  IsPortfolioEnabled: boolean;
}